09_RAG/python_helper_project/pyproject.toml
[project]
name = "python-helper-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pinecone>=6.0.1",
    "python-dotenv>=1.0.1",
]
